# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/gajis": {
            "get": {
                "tags": [
                    "Gaji"
                ],
                "summary": "List all gaji",
                "description": "Menampilkan daftar semua data gaji beserta karyawan",
                "operationId": "70b084ff3313ddf49fca39576bb1a9cf",
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan data gaji",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Gaji"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Gaji"
                ],
                "summary": "Create new gaji",
                "description": "Membuat data gaji baru",
                "operationId": "38547b46de2381846bd532a063a1cafe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "karyawan_id",
                                    "periode_gaji",
                                    "gaji_pokok",
                                    "total_jam_kerja",
                                    "bonus",
                                    "potongan",
                                    "total_gaji"
                                ],
                                "properties": {
                                    "karyawan_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "periode_gaji": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-28"
                                    },
                                    "gaji_pokok": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000000
                                    },
                                    "total-jam-kerja": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 160
                                    },
                                    "bonus": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500000
                                    },
                                    "potongan": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 200000
                                    },
                                    "total_gaji": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5300000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil membuat data gaji",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Gaji"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/gajis/{id}": {
            "get": {
                "tags": [
                    "Gaji"
                ],
                "summary": "Show specific gaji",
                "description": "Menampilkan detail data gaji berdasarkan ID",
                "operationId": "411447867f9a8bb19c04c55b0ffec8bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID data gaji",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail data gaji",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Gaji"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Gaji not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Gaji"
                ],
                "summary": "Update specific gaji",
                "description": "Memperbarui data gaji berdasarkan ID",
                "operationId": "a144be6d967d992f5d9117495318ea9f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID data gaji",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "karyawan_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "periode_gaji": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-28"
                                    },
                                    "gaji_pokok": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5500000
                                    },
                                    "total-jam-kerja": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 170
                                    },
                                    "bonus": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 600000
                                    },
                                    "potongan": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100000
                                    },
                                    "total_gaji": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 6000000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil memperbarui data gaji",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Gaji"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Gaji"
                ],
                "summary": "Delete specific gaji",
                "description": "Menghapus data gaji berdasarkan ID",
                "operationId": "086dd8aa864629a993d670a9206784c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID data gaji",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berhasil menghapus data gaji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/gajis/karyawan/{karyawan_id}": {
            "get": {
                "tags": [
                    "Gaji"
                ],
                "summary": "List gaji berdasarkan karyawan",
                "description": "Menampilkan daftar gaji berdasarkan ID karyawan",
                "operationId": "10ed44a85c9ba48384a707e0cbc68aa5",
                "parameters": [
                    {
                        "name": "karyawan_id",
                        "in": "path",
                        "description": "ID karyawan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan data gaji berdasarkan karyawan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Gaji"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Gaji": {
                "required": [
                    "karyawan_id",
                    "periode_gaji",
                    "gaji_pokok",
                    "total_jam_kerja",
                    "bonus",
                    "potongan",
                    "total_gaji"
                ],
                "properties": {
                    "id": {
                        "description": "ID Gaji",
                        "type": "integer",
                        "example": 1
                    },
                    "karyawan_id": {
                        "description": "ID Karyawan",
                        "type": "integer",
                        "example": 2
                    },
                    "periode_gaji": {
                        "description": "Periode Gaji",
                        "type": "string",
                        "format": "date",
                        "example": "2025-04-28"
                    },
                    "gaji_pokok": {
                        "description": "Gaji Pokok",
                        "type": "number",
                        "format": "float",
                        "example": 5000000
                    },
                    "total-jam-kerja": {
                        "description": "Total Jam Kerja",
                        "type": "number",
                        "format": "float",
                        "example": 160
                    },
                    "bonus": {
                        "description": "Bonus",
                        "type": "number",
                        "format": "float",
                        "example": 500000
                    },
                    "potongan": {
                        "description": "Potongan",
                        "type": "number",
                        "format": "float",
                        "example": 200000
                    },
                    "total_gaji": {
                        "description": "Total Gaji",
                        "type": "number",
                        "format": "float",
                        "example": 5300000
                    },
                    "created_at": {
                        "description": "Waktu Dibuat",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-28T08:00:00Z"
                    },
                    "updated_at": {
                        "description": "Waktu Diperbarui",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-28T09:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication using Laravel Sanctum",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "security": [
                {
                    "sanctum": []
                }
            ]
        }
    },
...
}